Conflicts:
	src/main/java/sepm/dsa/gui/MainMenuController.java
	src/main/java/sepm/dsa/gui/TradeSellToPlayerController.java

+++ src/main/java/sepm/dsa/service/TraderService.java
@@ -118,7 +118,7 @@ public interface TraderService {
 
     void addManualOffer(Trader trader, Offer offer);
 
-    void removeManualOffer(Trader trader, Offer offer, int amount);
+    void removeManualOffer(Trader trader, Offer offer, double amount);
 
     /**
      *
